refactor/rename schema files
