+++ mealie/routes/backup_routes.py
@@ -4,7 +4,7 @@ import shutil
 from core.config import BACKUP_DIR, TEMPLATE_DIR
 from db.db_setup import generate_session
 from fastapi import APIRouter, Depends, File, HTTPException, UploadFile
-from schema.backup_models import BackupJob, ImportJob, Imports, LocalBackup
+from schema.backup import BackupJob, ImportJob, Imports, LocalBackup
 from services.backups.exports import backup_all
 from services.backups.imports import ImportDatabase
 from sqlalchemy.orm.session import Session
