+++ mealie/schema/user.py
